Refactor Misc: move project python files to app
- Move requirements to requirements/ folder
- update README
- update TESTING
