+++ .travis.yml
@@ -16,8 +16,8 @@ addons:
 before_install:
   - ls
 install:
-  - pip install -r requirements.txt -t lib/
-  - pip install -r local_requirements.txt
+  - pip install -r requirements/requirements.txt -t lib/
+  - pip install -r requirements/local_requirements.txt
 
 before_script:
   - cd ..
