+++ TESTING
@@ -1,5 +1,5 @@
-We currently don't have any automated testing (see
-https://code.google.com/p/sympy/issues/detail?id=3578).  So everything has to
+We currently don't have a lot of automated testing (see
+https://github.com/sympy/sympy/issues/6677).  So most of the things has to
 be tested by hand.  Make sure to test the following:
 
 - All major web browsers.  Safari, Google Chrome, Firefox, and Internet
